EVG-18166: Refactor for better organization
