+++ src/components/NavBar/index.tsx
@@ -2,74 +2,21 @@ import { useState } from "react";
 import styled from "@emotion/styled";
 import IconButton from "@leafygreen-ui/icon-button";
 import { palette } from "@leafygreen-ui/palette";
-import { useLogWindowAnalytics } from "analytics";
 import DetailsOverlay from "components/DetailsOverlay";
 import Icon from "components/Icon";
 import PopoverButton from "components/PopoverButton";
-import SearchBar from "components/SearchBar";
+import Search from "components/Search";
 import ShortcutModal from "components/ShortcutModal";
 import { StyledLink } from "components/styles";
-import { CaseSensitivity, MatchType, SearchBarActions } from "constants/enums";
-import { QueryParams } from "constants/queryParams";
 import { navbarHeight, size } from "constants/tokens";
 import { useLogContext } from "context/LogContext";
-import { useFilterParam } from "hooks/useFilterParam";
-import { useQueryParam } from "hooks/useQueryParam";
-import { validateRegexp } from "utils/validators";
-import SearchResults from "./SearchResults";
 import UploadLink from "./UploadLink";
 
 const { gray, white } = palette;
 
 const NavBar: React.FC = () => {
   const [open, setOpen] = useState(false);
-  const [filters, setFilters] = useFilterParam();
-  const [highlights, setHighlights] = useQueryParam<string[]>(
-    QueryParams.Highlights,
-    []
-  );
-  const { hasLogs, clearLogs, setSearch, searchState, paginate } =
-    useLogContext();
-  const { sendEvent } = useLogWindowAnalytics();
-  const { hasSearch } = searchState;
-  const handleSearch = (selected: string, value: string) => {
-    switch (selected) {
-      case SearchBarActions.Search:
-        setSearch(value);
-        break;
-      case SearchBarActions.Filter:
-        if (!filters.some((f) => f.name === value)) {
-          setSearch("");
-          setFilters([
-            ...filters,
-            {
-              name: value,
-              caseSensitive: CaseSensitivity.Insensitive,
-              matchType: MatchType.Exact,
-              visible: true,
-            },
-          ]);
-          sendEvent({ name: "Added Filter", filterExpression: value });
-        }
-        break;
-      case SearchBarActions.Highlight:
-        if (!highlights.includes(value)) {
-          setSearch("");
-          setHighlights([...highlights, value]);
-          sendEvent({ name: "Added Highlight", highlightExpression: value });
-        }
-        break;
-      default:
-        throw new Error("Invalid search action");
-    }
-  };
-
-  const handleOnChange = (selected: string, value: string) => {
-    if (selected === SearchBarActions.Search) {
-      setSearch(value);
-      sendEvent({ name: "Applied Search", searchExpression: value });
-    }
-  };
+  const { hasLogs, clearLogs } = useLogContext();
 
   return (
     <Container>
@@ -81,16 +28,7 @@ const NavBar: React.FC = () => {
           </StyledLink>
           <UploadLink clearLogs={clearLogs} hasLogs={hasLogs} />
         </LinkContainer>
-        <StyledSearchBar
-          disabled={!hasLogs}
-          onChange={handleOnChange}
-          onSubmit={handleSearch}
-          validator={validateRegexp}
-          validatorMessage="Invalid Regular Expression"
-        />
-        {hasSearch && (
-          <SearchResults paginate={paginate} searchState={searchState} />
-        )}
+        <Search />
       </FlexContainer>
 
       <FlexContainer>
@@ -139,11 +77,6 @@ const LinkContainer = styled.div`
   display: flex;
   margin-right: ${size.l};
   gap: ${size.l};
-`;
-
-const StyledSearchBar = styled(SearchBar)`
-  width: 60vw;
-  margin-left: ${size.m};
 `;
 
 const StyledButton = styled(PopoverButton)`
