+++ src/components/Search/SearchBar/SearchBar.test.tsx
