+++ src/components/Search/SearchResults/SearchCount.tsx
