+++ src/components/SidePanel/NavGroup/BaseNavGroup/index.tsx
@@ -1,0 +1,70 @@
+import { PropsWithChildren, ReactNode } from "react";
+import styled from "@emotion/styled";
+import Badge, { Variant } from "@leafygreen-ui/badge";
+import { palette } from "@leafygreen-ui/palette";
+import { SideNavGroup } from "@leafygreen-ui/side-nav";
+import { Body } from "@leafygreen-ui/typography";
+import Icon, { glyphs } from "components/Icon";
+import { size } from "constants/tokens";
+
+const { green } = palette;
+
+interface BaseNavGroupProps<T> {
+  ["data-cy"]: string;
+  children: ReactNode;
+  glyph: keyof typeof glyphs;
+  items: T[];
+  navGroupTitle: string;
+  defaultMessage: string;
+}
+
+const BaseNavGroup = <T extends {}>({
+  "data-cy": dataCy,
+  children,
+  glyph,
+  items,
+  navGroupTitle,
+  defaultMessage,
+}: PropsWithChildren<BaseNavGroupProps<T>>) => (
+  <StyledSideNavGroup
+    glyph={<Icon fill={green.dark2} glyph={glyph} />}
+    hasActiveItem={items.length > 0}
+    header={
+      <NavGroupHeader data-cy={`${dataCy}-nav-group-header`}>
+        <NavGroupTitle>{navGroupTitle}</NavGroupTitle>
+        <Badge variant={Variant.Green}>{items.length}</Badge>
+      </NavGroupHeader>
+    }
+  >
+    {items.length ? (
+      children
+    ) : (
+      <SectionWrapper data-cy={`${dataCy}-default-message`}>
+        <Body>{defaultMessage}</Body>
+      </SectionWrapper>
+    )}
+  </StyledSideNavGroup>
+);
+
+// @ts-expect-error
+const StyledSideNavGroup = styled(SideNavGroup)`
+  > div {
+    padding: 0;
+  }
+`;
+
+const NavGroupHeader = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+const NavGroupTitle = styled.div`
+  margin-right: ${size.xxs};
+`;
+
+const SectionWrapper = styled.div`
+  margin-top: ${size.xs};
+  margin-bottom: ${size.s};
+`;
+
+export default BaseNavGroup;
