+++ src/components/SidePanel/NavGroup/index.tsx
@@ -1,0 +1,5 @@
+import ExpandedNavGroup from "./ExpandedNavGroup";
+import FilterNavGroup from "./FilterNavGroup";
+import HighlightNavGroup from "./HighlightNavGroup";
+
+export { ExpandedNavGroup, FilterNavGroup, HighlightNavGroup };
